requirements.txt
python-dotenv
Pillow
python-telegram-bot
telethon

